+++ packages/jp.co.hololab.dnn.objectdetection/Documentation/CHANGES.md
@@ -1,5 +1,13 @@
 # CHANGES
 
+## [1.0.2] - 2024-XX-XX
+
+- Update Unity Sentis to 1.5 from 1.4.
+- Update Unity Required Version to 2023.2 (Unity 6) or later.
+- Remove Constructor to Load Model from ONNX at Runtime.
+- Add Constructor to Load Model from Sentis (*.sentis) at Runtime.
+- Add Constructor to Load Model from Stream at Runtime.
+
 ## [1.0.1] - 2024-04-10
 
 - Update Unity Sentis to 1.4 from 1.3.